+++ test/enrollment_repo_test.rb
@@ -27,23 +27,9 @@ class EnrollmentRepositoryTest < MiniTest::Test
     assert_respond_to(@er, :find_by_name)
   end
 
-  def test_responds_to_create_enrollment
-    assert_respond_to(@er, :create_enrollment)
-  end
-
-  def test_responds_to_process_data
-    assert_respond_to(@er, :process_data)
-  end
-
-  def test_create_enrollment_adds_enrollment_object_to_data
-    enrollment = @er.create_enrollment(@enrollment_data)
-    assert_equal 1, @er.data.length
-    assert_instance_of Enrollment, @er.data[0]
-  end
-
   def test_find_by_name_returns_enrollment
     assert_equal 0, @er.data.length
-    enrollment = @er.create_enrollment(@enrollment_data)
+    enrollment = EnrollmentBuilder.new(@er).create_enrollment(@enrollment_data)
     assert_equal 1, @er.data.length
     enrollment = @er.find_by_name("DISTRICT 1")
     assert_equal enrollment, @er.data[0]
