+++ test/test_helper.rb
@@ -10,6 +10,7 @@ require_relative '../lib/csv_parser'
 require_relative '../lib/district_repository'
 require_relative '../lib/district'
 require_relative '../lib/enrollment_repository'
+require_relative '../lib/builders/enrollment_builder'
 require_relative '../lib/enrollment'
 require_relative '../lib/statewide_testing_repository'
 require_relative '../lib/statewide_test'
